Extracted Queryable comparison operator code.
